Removed old sidenav standardising component
